Move mesh validation tests to `test_dataobject_filters.py`
